refactor: [#1320] move code from axum-http-tracker-server pkg to http-protocol pkg
