--- packages/axum-http-tracker-server/src/v1/handlers/common/mod.rs
@@ -1,3 +1,0 @@
-//! Common logic for HTTP handlers.
-pub mod auth;
-pub mod peer_ip;
