+++ packages/axum-http-tracker-server/src/v1/handlers/mod.rs
@@ -1,5 +1,4 @@
 //! Axum [`handlers`](axum#handlers) for the HTTP server.
 pub mod announce;
-pub mod common;
 pub mod health_check;
 pub mod scrape;
