+++ packages/http-protocol/src/v1/mod.rs
@@ -1,3 +1,4 @@
+pub mod auth;
 pub mod query;
 pub mod requests;
 pub mod responses;
