+++ packages/http-protocol/src/v1/responses/error.rs
@@ -13,6 +13,7 @@
 //! > code.
 use serde::Serialize;
 
+use crate::v1::auth;
 use crate::v1::services::peer_ip_resolver::PeerIpResolutionError;
 
 /// `Error` response for the HTTP tracker.
@@ -44,6 +45,14 @@ impl Error {
     #[must_use]
     pub fn write(&self) -> String {
         serde_bencode::to_string(&self).unwrap()
+    }
+}
+
+impl From<auth::Error> for Error {
+    fn from(err: auth::Error) -> Self {
+        Self {
+            failure_reason: format!("Tracker authentication error: {err}"),
+        }
     }
 }
 
@@ -89,8 +98,11 @@ impl From<bittorrent_tracker_core::authentication::Error> for Error {
 
 #[cfg(test)]
 mod tests {
+    use std::panic::Location;
 
     use super::Error;
+    use crate::v1::responses;
+    use crate::v1::services::peer_ip_resolver::PeerIpResolutionError;
 
     #[test]
     fn http_tracker_errors_can_be_bencoded() {
@@ -99,5 +111,21 @@ mod tests {
         };
 
         assert_eq!(err.write(), "d14:failure reason13:error messagee"); // cspell:disable-line
+    }
+
+    fn assert_error_response(error: &responses::error::Error, error_message: &str) {
+        assert!(
+            error.failure_reason.contains(error_message),
+            "Error response does not contain message: '{error_message}'. Error: {error:?}"
+        );
+    }
+
+    #[test]
+    fn it_should_map_a_peer_ip_resolution_error_into_an_error_response() {
+        let response = responses::error::Error::from(PeerIpResolutionError::MissingRightMostXForwardedForIp {
+            location: Location::caller(),
+        });
+
+        assert_error_response(&response, "Error resolving peer IP");
     }
 }
